challenges/signum_paranoia/src/requirements.txt
click==8.0.3
Flask==2.0.2
Flask-Limiter==2.1.3
itsdangerous==2.0.1
Jinja2==3.0.3
limits==2.3.3
MarkupSafe==2.0.1
Werkzeug==2.0.2
gunicorn==20.1.0
python-dotenv==0.19.2